small warning fixes

+++ RasterizerDemo/InputLayoutD3D11.cpp
@@ -3,7 +3,7 @@
 void InputLayoutD3D11::AddInputElement(const char* semanticName, DXGI_FORMAT format, size_t inputSlot)
 {
 	D3D11_INPUT_ELEMENT_DESC inputDesc =
-	{ semanticName, 0, format, 0, inputSlot, D3D11_INPUT_PER_VERTEX_DATA, 0 };
+	{ semanticName, 0, format, 0, (UINT)inputSlot, D3D11_INPUT_PER_VERTEX_DATA, 0 };
 
 
     this->elements.push_back(inputDesc);
@@ -12,7 +12,7 @@ void InputLayoutD3D11::AddInputElement(const char* semanticName, DXGI_FORMAT for
 
 void InputLayoutD3D11::FinalizeInputLayout(ID3D11Device* device, const void* vsDataPtr, size_t vsDataSize)
 {
-	if (FAILED(device->CreateInputLayout(this->elements.data(), this->elements.size(), vsDataPtr, vsDataSize, this->inputLayout.GetAddressOf())))
+	if (FAILED(device->CreateInputLayout(this->elements.data(), (UINT)this->elements.size(), vsDataPtr, vsDataSize, this->inputLayout.GetAddressOf())))
 	{
 		std::cerr << "Failed to create Input Layout!" << std::endl;
 		throw std::runtime_error("Failed to create Input Layout!");
